refactor(crawler): fuse extract_content into extraction module (#442)
Move the content-extraction pipeline (clean -> selector -> adaptive repair ->
readability/fallback) out of the discovery.rs God Object into
application/extraction.rs, alongside extract_with_selector. It now calls the
local extract_with_selector and asset_download::download_assets_if_enabled
directly.

discovery.rs re-exports extract_content so scrape_single_url_for_tui (which
delegates to it), the crawler facade and lib.rs resolve unchanged; the
extract_content unit tests stay in discovery.rs and exercise the re-export.
Verified under default features and --features adaptive-selectors. Pure move,
no behavior change.
